complete runnable train/test script
