--- run/__init__.py
@@ -1,4 +1,0 @@
-import model as model_modules
-import data_loader as data_loader_modules
-import metric as metric_modules
-import loss_function as loss_fn_modules
