+++ utils/workspace.py
@@ -7,14 +7,19 @@ from .file_utils import ensure_dir
 
 
 class Workspace():
-    def __init__(self, dir, model, loss_fn, metric, optimizer=None):
+    def __init__(self, device, dir, model, loss_fn, metric, optimizer=None):
+        self.device = device
         self.dir = dir
         self.model = model
         self.loss_fn = loss_fn
         self.metric = metric
         self.optimizer = optimizer
 
-        ensure_dir(self.dir)
+        path = Path(self.dir)
+
+        if not path.exists():
+            # Training case
+            ensure_dir(self.dir)
 
     def _get_checkpoint_path(self, epoch):
         return Path(self.dir, "checkpoint_{}.pt".format(epoch))
@@ -46,21 +51,23 @@ class Workspace():
         self._save(path, accessories)
 
     def _load(self, path):
-        checkpoint = torch.load(path)
+        checkpoint = torch.load(path, map_location=self.device)
 
         self.model.load_state_dict(checkpoint.pop('model_state_dict'))
         self.loss_fn = checkpoint.pop('loss_fn')
         self.metric = checkpoint.pop('metric')
 
-        if self.optimizer:
+        if self.optimizer is not None:
             self.optimizer.load_state_dict(checkpoint.pop('optimizer_state_dict'))
 
         return checkpoint
 
     def load_checkpoint(self, epoch):
         path = self._get_checkpoint_path(epoch)
+        print(f"loading {path}")
         return self._load(path)
 
     def load_best_model(self):
         path = self._get_best_model_path()
+        print(f"loading {path}")
         return self._load(path)
